remove compiler/ssair/driver.jl
I don't think that the extra layer of indirection is needed.
